Move telemetry utils to their own file Rename telemetry_utils to telemetry Refactor parse_radio_block function decoupling from state
